'tests' folder moved from a package to project root. pytes.ini added.
