+++ pytest.ini
@@ -1,0 +1,5 @@
+
+[pytest]
+minversion = 6.0
+addopts = -ra -q -rf
+testpaths = tests
