+++ tests/core/batch.py
