+++ tests/loaders/pandas.py
@@ -13,33 +13,41 @@
 #     See the License for the specific language governing permissions and
 #     limitations under the License.
 
-
-from os import path
 import pytest
-from witness.extractors.pandas import PandasFeatherExtractor
+from os import path
+from witness import Batch
+from witness.loaders.pandas import PandasFeatherLoader, PandasExcelLoader
+
+from tests.core.batch import calibration_meta, calibration_data
 
 xfail = pytest.mark.xfail
 parametrize = pytest.mark.parametrize
 
+
 # region mock
-mock_dir = path.abspath('../../../mock')
+mock_dir = path.abspath('../../mock')
 files_dir = f'{mock_dir}/files'
+calibration_batch = Batch(calibration_data, calibration_meta)
+
 mock_params = [
-    (PandasFeatherExtractor, f'{files_dir}/feather_dump')
+    (PandasFeatherLoader, f'{files_dir}/feather_dump', calibration_batch),
+    (PandasExcelLoader, f'{files_dir}/excel_dump.xlsx', calibration_batch)
 ]
+
+
 # endregion mock
 
 
-@parametrize('extractor, uri', mock_params)
-def test_create(extractor, uri):
-    new_extractor = extractor(uri=uri)
-    return new_extractor.uri
+@parametrize('loader, uri, batch', mock_params)
+def test_prepare(loader, uri, batch):
+    new_loader = loader(uri)
+    new_loader.prepare(batch)
 
 
-@parametrize('extractor, uri', mock_params)
-def test_extract(extractor, uri):
-    new_extractor = extractor(uri=uri)
-    new_extractor.extract()
+@parametrize('loader, uri, batch', mock_params)
+def test_load(loader, uri, batch):
+    new_loader = loader(uri)
+    new_loader.prepare(batch).load()
 
 
 if __name__ == '__main__':
